refactor(agent): fold AgentConfig host imports through @agent/runtime
Address the #10712 P1 that the definition barrel was a 1-for-1
baseline edge swap: re-export the host-used AgentConfig contract from
the already-approved @agent/runtime door, migrate the 14 host imports,
delete src/agent/core/definition/index.ts, and remove the
@agent/core/definition/AgentConfig specifier from cli/desktop/extension
without adding a replacement. The host baseline shrinks by 3
(cli 13->12, desktop 11->10, extension 14->13).
