+++ docs/proposals/2026-08-04-agent-sdk-readiness-review.md
@@ -126,11 +126,8 @@ extension's model-handler / text-enhancement imports and its
 `review/`+`agentCreator/`+`goal/` composition imports (see §4).
 
 The invariant to hold while doing this (per CLAUDE.md): **never widen a
-baseline.** A promotion into the Tier-1 barrel should _shrink_ the matching
-host list where a cluster has more than one deep specifier; where a cluster
-already has one deep specifier per host, the promotion may swap that deep
-specifier 1-for-1 for the barrel door instead of a numeric shrink. Either way,
-never add a new export without a consumer.
+baseline.** Each promotion into the Tier-1 barrel should _shrink_ the matching
+host list, not add a new export without a consumer.
 
 ---
 
@@ -198,13 +195,12 @@ These are small and independent of the strategic Tier-1 program:
    `src/agent/followUp/index.ts` barrel (`submitFollowUp`,
    `presentFollowUpResult`, `notifyFollowUpSent`, and the
    `SubmitFollowUpResult`/`FollowUpQueueInput`/`FollowUpRecoveryLease` types).
-   The final `core/definition/AgentConfig` cluster folded behind a new curated
-   `src/agent/core/definition/index.ts` barrel (`AgentConfigSchema`,
+   The final `core/definition/AgentConfig` cluster folded through the
+   already-approved `@agent/runtime` door (`AgentConfigSchema`,
    `AgentConfig`, `AgentConfigPayload`), migrating every host off the deep
-   `@agent/core/definition/AgentConfig` path. Because that cluster already had
-   one deep specifier per host, the baseline change is a 1-for-1 door swap
-   (cli 13→13, desktop 11→11, extension 14→14, total 45→45) with no widening,
-   completing #10024.
+   `@agent/core/definition/AgentConfig` path and deleting that specifier from
+   each host list (cli 13→12, desktop 11→10, extension 14→13; 38→35 across the
+   three host lists, 45→42 with the SDK package unchanged), completing #10024.
 
 2. **Stabilize the withheld interaction contract.** The `HostInteractions`
    docstring (`index.ts:42-47`) and the hard-deny `requestRetry`
@@ -237,5 +233,5 @@ surface; see §5.1 and review `pullrequestreview-4918028384`. #10531 continued
 the fold-in behind the pre-existing module-level `@agent/storage` barrel, and
 #10650 folded the `followUp` cluster behind a new curated
 `src/agent/followUp/index.ts` barrel. #10712 folded the
-`core/definition/AgentConfig` cluster behind a new curated
-`src/agent/core/definition/index.ts` barrel; #10024 has no remaining clusters.
+`core/definition/AgentConfig` cluster through the already-approved
+`@agent/runtime` door; #10024 has no remaining clusters.
