+++ packages/cli/src/runtime/toolUseResumeData.ts
@@ -1,9 +1,9 @@
 import { getExecutionStore } from '@agent/storage';
 import {
   retrieveSessionResumeData,
+  type AgentConfig,
   type ToolUseResumeData,
 } from '@agent/runtime';
-import type { AgentConfig } from '@agent/core/definition';
 import { createLog } from '@logger/logUtils';
 import type { ExecutionId } from '@shared/schemas';
 import { AgentCategory } from '@shared/schemas';
