+++ packages/extension/src/commands/agent/executeCommand.ts
@@ -3,8 +3,11 @@ import * as vscode from 'vscode';
 import { z, ZodError } from 'zod';
 
 // Local imports
-import { ModelHandlerCompatibilityKeySchema, runAgent } from '@agent/runtime';
-import { AgentConfigSchema } from '@agent/core/definition';
+import {
+  AgentConfigSchema,
+  ModelHandlerCompatibilityKeySchema,
+  runAgent,
+} from '@agent/runtime';
 import { openFinalOutputIfAvailable } from '@frontend/agents/finalOutputOpener';
 import { createLog } from '@logger/logUtils';
 import type { ExecutionId } from '@shared/schemas';
